chore: Split into two test cases
